Fit a regression model based on response type

S3 generic that dispatches to the appropriate model fitting function based on the class of the response variable.

autoreg(
  response,
  data,
  x_var,
  id = NULL,
  strata = "",
  family = NULL,
  offset = NULL,
  corstr = "independence"
)

Arguments

response

The response variable (used for S3 dispatch).

data

A data frame containing the model data.

x_var

Character string of the predictor variable name.

id

Optional subject ID for GEE models.

strata

Optional stratification variable.

family

Optional family for GLM models.

offset

Optional offset term.

corstr

Correlation structure for GEE models (default "independence").

Value

A fitted model object.