Tumour size change over time Longitudinal changes in tumour size since baseline for patients by changes in ctDNA status (clearance, decrease or increase) since baseline.

Tumour size change over time

Longitudinal changes in tumour size since baseline for patients by changes in ctDNA status (clearance, decrease or increase) since baseline.

data('ctDNA')

Format

A data frame with 270 rows and 5 variables:

id

Patient ID

cohort

Study Cohort

ctdna_status

Change in ctDNA since baseline

time

Number of weeks on treatment

size_change

Percentage change in tumour measurement

Source

https://www.nature.com/articles/s43018-020-0096-5