Additional parameters passed to ggkmcif2

This section documents the additional parameters for ggkmcif2.

Arguments

HR

boolean to specify if you want hazard ratios included in the plot

HR_pval

boolean to specify if you want HR p-values in the plot

conf.type

One of "none"(the default), "plain", "log" , "log-log" or "logit". Only enough of the string to uniquely identify it is necessary. The first option causes confidence intervals not to be generated. The second causes the standard intervals curve +- k *se(curve), where k is determined from conf.int. The log option calculates intervals based on the cumulative hazard or log(survival). The log-log option bases the intervals on the log hazard or log(-log(survival)), and the logit option on log(survival/(1-survival)).

table.height

Relative height of risk table (0-1)

main

String corresponding to main title. When NULL uses Kaplan-Meier Plot s, and "Cumulative Incidence Plot for CIF"

stratalabs

string corresponding to the labels of the covariate, when NULL will use the levels of the covariate

strataname

String of the covariate name default is nicename(cov)

stratalabs.table

String corresponding to the levels of the covariate for the number at risk table, when NULL will use the levels of the covariate. Can use a string of "-" when the labels are long

strataname.table

String of the covariate name for the number at risk table default is nicename(cov

median.text

boolean to specify if you want the median values added to the legend (or as added text if there are no covariates), for KM only

median.lines

boolean to specify if you want the median values added as lines to the plot, for KM only

median.CI

boolean to specify if you want the 95\ interval with the median text (Only for KM)

set.time.text

string for the text to add survival at a specified time (eg. year OS)

set.time.line

boolean to specify if you want the survival added as lines to the plot at a specified point

set.time

Numeric values of the specific time of interest, default is 5 (Multiple values can be entered)

set.time.CI

boolean to specify if you want the 95\ interval with the set time text

censor.marks

logical value. If TRUE, includes censor marks (only for KM curves)

censor.size

size of censor marks, default is 3

censor.stroke

stroke of censor marks, default is 1.5

fsize

font size

nsize

font size for numbers in the numbers at risk table

lsize

line size

psize

size of the pvalue

median.size

size of the median text (Only when there are no covariates)

median.pos

vector of length 2 corresponding to the median position (Only when there are no covariates)

median.lsize

line size of the median lines

set.size

size of the survival at a set time text (Only when there are no covariates)

set.pos

vector of length 2 corresponding to the survival at a set point position (Only when there are no covariates)

set.lsize

line size of the survival at set points

ylim

vector of length 2 corresponding to limits of y-axis. Default to NULL

linetype

vector of line types; default is solid for all lines

xlim

vector of length 2 corresponding to limits of x-axis. Default to NULL

legend.pos

Can be either a string corresponding to the legend position ("left","top", "right", "bottom", "none") or a vector of length 2 corresponding to the legend position (uses normalized units (ie the c(0.5,0.5) is the middle of the plot))

pval.pos

vector of length 2 corresponding to the p-value position

event

String specifying if the event should be mapped to the colour, or linetype when plotting both events to colour = "col", line type

flip.CIF

boolean to flip the CIF curve to start at 1

cut

numeric value indicating where to divide a continuous covariate (default is the median)

eventlabs

String corresponding to the event type names

event.name

String corresponding to the label of the event types

Numbers_at_risk_text

String for the label of the number at risk

HR.digits

Number of digits printed of the hazard ratio

HR.pval.digits

Number of digits printed of the hazard ratio pvalue

pval.digits

Number of digits printed of the Gray's/log rank pvalue

median.digits

Number of digits printed of the median pvalue

set.time.digits

Number of digits printed of the probability at a specified time

print.n.missing

Logical, should the number of missing be shown !Needs to be checked

returns

Logical, if TRUE a list contain the plot and at risk table is returned