Combine two table columns into a single column with levels of one nested within levels of the other.

This function accepts a data frame (via the data argument) and combines two columns into a single column with values from the head_col serving as headers and values of the to_col displayed underneath each header. The resulting table is then passed to outTable for printing and output, to use the grouped table as a data frame specify tableOnly=TRUE. By default the headers will be bolded and the remaining values indented.

nestTable(
  data,
  head_col,
  to_col,
  colHeader = "",
  caption = NULL,
  indent = TRUE,
  boldheaders = TRUE,
  hdr_prefix = "",
  hdr_suffix = "",
  digits = getOption("reportRmd.digits", 2),
  tableOnly = FALSE,
  fontsize
)

Arguments

data

dataframe

head_col

character value specifying the column name with the headers

to_col

character value specifying the column name to add the headers into

colHeader

character with the desired name of the first column. The default is to leave this empty for output or, for table only output to use the column name 'col1'.

caption

table caption

indent

Boolean should the original values in the to_col be indented

boldheaders

Boolean should the header column values be bolded

hdr_prefix

character value that will prefix headers

hdr_suffix

character value that will suffix headers

digits

number of digits to round numeric columns to, wither a single number or a vector corresponding to the number of numeric columns

tableOnly

boolean indicating if the table should be formatted for printing or returned as a data frame

fontsize

PDF/HTML output only, manually set the table fontsize

Value

A character vector of the table source code, unless tableOnly=TRUE in which case a data frame is returned

Details

Note that it is possible to combine multiple tables (more than two) with this function.

Examples

## Investigate models to predict baseline ctDNA and tumour size and display together
## (not clinically useful!)
data(pembrolizumab)
fit1 <- lm(baseline_ctdna~age+l_size+pdl1,data=pembrolizumab)
m1 <- rm_mvsum(fit1,tableOnly=TRUE)
m1$Response = 'ctDNA'
fit2 <- lm(l_size~age+baseline_ctdna+pdl1,data=pembrolizumab)
m2 <- rm_mvsum(fit2,tableOnly=TRUE)
m2$Response = 'Tumour Size'
nestTable(rbind(m1,m2),head_col='Response',to_col='Covariate')
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Estimate(95%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> VIF </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">ctDNA</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Age at study entry </td>
#>    <td style="text-align:right;"> 0.99 (-10.15, 12.13) </td>
#>    <td style="text-align:right;"> 0.86 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.03 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Target lesion size at baseline </td>
#>    <td style="text-align:right;"> 1.00 (-1.43, 3.43) </td>
#>    <td style="text-align:right;"> 0.41 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.06 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> PD L1 percent </td>
#>    <td style="text-align:right;"> -3.08 (-8.00, 1.83) </td>
#>    <td style="text-align:right;"> 0.22 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.05 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Tumour Size</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Age at study entry </td>
#>    <td style="text-align:right;"> -0.69 (-1.64, 0.26) </td>
#>    <td style="text-align:right;"> 0.15 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.01 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> PD L1 percent </td>
#>    <td style="text-align:right;"> -0.38 (-0.80, 0.04) </td>
#>    <td style="text-align:right;"> 0.076 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.03 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Baseline ctDNA </td>
#>    <td style="text-align:right;"> 7.5e-03 (-0.01, 0.03) </td>
#>    <td style="text-align:right;"> 0.41 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 1.02 </td>
#>   </tr>
#> </tbody>
#> </table>