Display survival rates and events for specified times

This is a wrapper for the survfit function to output a tidy display for reporting. Either Kaplan Meier or Cox Proportional Hazards models may be used to estimate the survival probabilities.

rm_survtime(
  data,
  time,
  status,
  covs = NULL,
  strata = NULL,
  type = "KM",
  survtimes,
  survtimeunit,
  strata.prefix = NULL,
  survtimesLbls = NULL,
  showCols = c("At Risk", "Events", "Censored"),
  CIwidth = 0.95,
  conf.type = "log",
  na.action = "na.omit",
  showCounts = TRUE,
  digits = getOption("reportRmd.digits", 2),
  caption = NULL,
  tableOnly = FALSE,
  fontsize
)

Arguments

data

data frame containing survival data

time

string indicating survival time variable

status

string indicating event status variable

covs

character vector with the names of variables to adjust for in coxph fit

strata

string indicating the variable to group observations by. If this is left as NULL (the default) then event counts and survival rates are provided for the entire cohort.

type

survival function, if no covs are specified defaults to Kaplan-Meier, otherwise the Cox PH model is fit. Use type='PH' to fit a Cox PH model with no covariates.

survtimes

numeric vector specifying when survival probabilities should be calculated.

survtimeunit

unit of time to suffix to the time column label if survival probabilities are requested, should be plural

strata.prefix

character value describing the grouping variable

survtimesLbls

if supplied, a vector the same length as survtimes with descriptions (useful for displaying years with data provided in months)

showCols

character vector specifying which of the optional columns to display, defaults to c('At Risk','Events','Censored')

CIwidth

width of the survival probabilities, default is 95%

conf.type

type of confidence interval see survival::survfit for details. Default is 'log'.

na.action

default is to omit missing values, but can be set to throw and error using na.action='na.fail'

showCounts

boolean indicating if the at risk, events and censored columns should be output, default is TRUE

digits

the number of digits in the survival rate, default is 2.

caption

table caption for markdown output

tableOnly

should a dataframe or a formatted object be returned

fontsize

PDF/HTML output only, manually set the table fontsize

Value

A character vector of the survival table source code, unless tableOnly=TRUE in which case a data frame is returned

Details

If covariates are supplied then a Cox proportional hazards model is fit for the entire cohort and each strata. Otherwise the default is for Kaplan-Meier estimates. Setting type = 'PH' will force a proportional hazards model.

See also

Examples

# Kaplan-Mieir survival probabilities with time displayed in years
data("pembrolizumab")
rm_survtime(data=pembrolizumab,time='os_time',status='os_status',
strata="cohort",type='KM',survtimes=seq(12,72,12),
survtimesLbls=seq(1,6,1),
survtimeunit='years')
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;"> Time (years) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> At Risk </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Events </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Censored </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Survival Rate (95% CI) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Overall </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 48 </td>
#>    <td style="text-align:right;"> 44 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.53 (0.44, 0.64) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 24 </td>
#>    <td style="text-align:right;"> 17 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0.33 (0.24, 0.44) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 21 </td>
#>    <td style="text-align:right;"> 0.28 (0.20, 0.40) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> A </td>
#>    <td style="text-align:right;"> 16 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 10 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.38 (0.20, 0.71) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0.23 (0.09, 0.59) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.23 (0.09, 0.59) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> B </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 5 </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.32 (0.16, 0.64) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.06 (9.6e-03, 0.42) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> C </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.61 (0.42, 0.88) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.44 (0.27, 0.74) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.31 (0.15, 0.64) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> D </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.67 (0.45, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.67 (0.45, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.67 (0.45, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> E </td>
#>    <td style="text-align:right;"> 30 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.63 (0.48, 0.83) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 9 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.34 (0.20, 0.57) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 0.29 (0.15, 0.53) </td>
#>   </tr>
#> </tbody>
#> </table>

# Cox Proportional Hazards survivial probabilities
rm_survtime(data=pembrolizumab,time='os_time',status='os_status',
strata="cohort",type='PH',survtimes=seq(12,72,12),survtimeunit='months')
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;"> Time (months) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> At Risk </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Events </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Censored </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Survival Rate (95% CI) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Overall </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 48 </td>
#>    <td style="text-align:right;"> 44 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.53 (0.44, 0.64) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 24 </td>
#>    <td style="text-align:right;"> 17 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0.33 (0.25, 0.45) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 21 </td>
#>    <td style="text-align:right;"> 0.28 (0.20, 0.40) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> A </td>
#>    <td style="text-align:right;"> 16 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 10 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.39 (0.22, 0.72) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0.26 (0.11, 0.60) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.26 (0.11, 0.60) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> B </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 5 </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.34 (0.18, 0.65) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.09 (0.02, 0.41) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> C </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.62 (0.44, 0.89) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.46 (0.28, 0.75) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.33 (0.17, 0.65) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> D </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.68 (0.47, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.68 (0.47, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.68 (0.47, 0.99) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> E </td>
#>    <td style="text-align:right;"> 30 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.64 (0.49, 0.83) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 9 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.35 (0.22, 0.58) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 0.30 (0.17, 0.54) </td>
#>   </tr>
#> </tbody>
#> </table>

# Cox Proportional Hazards survivial probabilities controlling for age
rm_survtime(data=pembrolizumab,time='os_time',status='os_status',
covs='age',strata="cohort",survtimes=seq(12,72,12),survtimeunit='months')
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;"> Time (months) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> At Risk </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Events </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Censored </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Survival Rate (95% CI) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Overall </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 48 </td>
#>    <td style="text-align:right;"> 44 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.54 (0.44, 0.65) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 24 </td>
#>    <td style="text-align:right;"> 17 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0.33 (0.25, 0.45) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 21 </td>
#>    <td style="text-align:right;"> 0.28 (0.20, 0.40) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> A </td>
#>    <td style="text-align:right;"> 16 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 10 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.39 (0.22, 0.72) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0.26 (0.11, 0.60) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.26 (0.11, 0.60) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> B </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 5 </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.33 (0.17, 0.65) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.07 (0.01, 0.45) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> C </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.65 (0.46, 0.92) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.47 (0.29, 0.78) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.32 (0.16, 0.68) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> D </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0.70 (0.48, 1.00) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 0.70 (0.48, 1.00) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0.70 (0.48, 1.00) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> E </td>
#>    <td style="text-align:right;"> 30 </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 12 </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;"> 11 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.64 (0.49, 0.84) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 24 </td>
#>    <td style="text-align:right;"> 9 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0.35 (0.22, 0.58) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> 36 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 0.30 (0.16, 0.54) </td>
#>   </tr>
#> </tbody>
#> </table>