Output several univariate models nicely in a single table

#'A table with the model parameters from running separate univariate models on each covariate. For factors with more than two levels a Global p-value is returned.

rm_uvsum(
  response,
  covs,
  data,
  digits = getOption("reportRmd.digits", 2),
  covTitle = "",
  caption = NULL,
  tableOnly = FALSE,
  removeInf = FALSE,
  p.adjust = "none",
  unformattedp = FALSE,
  whichp = c("levels", "global", "both"),
  chunk_label,
  gee = FALSE,
  id = NULL,
  corstr = NULL,
  family = NULL,
  type = NULL,
  offset = NULL,
  strata = 1,
  nicenames = TRUE,
  showN = TRUE,
  showEvent = TRUE,
  CIwidth = 0.95,
  reflevel = NULL,
  returnModels = FALSE,
  fontsize,
  forceWald = FALSE
)

Arguments

response

string vector with name of response

covs

character vector with the names of columns to fit univariate models to

data

dataframe containing data

digits

number of digits to round estimates and CI to. Does not affect p-values.

covTitle

character with the names of the covariate (predictor) column. The default is to leave this empty for output or, for table only output to use the column name 'Covariate'.

caption

character containing table caption (default is no caption)

tableOnly

boolean indicating if unformatted table should be returned

removeInf

boolean indicating if infinite estimates should be removed from the table

p.adjust

p-adjustments to be performed. Uses the p.adjust function from base R

unformattedp

boolean indicating if you would like the p-value to be returned unformatted (ie not rounded or prefixed with '<'). Should be used in conjunction with the digits argument.

whichp

string indicating whether you want to display p-values for levels within categorical data ("levels"), global p values ("global"), or both ("both"). Irrelevant for continuous predictors.

chunk_label

only used if output is to Word to allow cross-referencing

gee

boolean indicating if gee models should be fit to account for correlated observations. If TRUE then the id argument must specify the column in the data which indicates the correlated clusters.

id

character vector which identifies clusters. Only used for geeglm

corstr

character string specifying the correlation structure. Only used for geeglm. The following are permitted: '"independence"', '"exchangeable"', '"ar1"', '"unstructured"' and '"userdefined"'

family

description of the error distribution and link function to be used in the model. Only used for geeglm

type

string indicating the type of univariate model to fit. The function will try and guess what type you want based on your response. If you want to override this you can manually specify the type. Options include "linear", "logistic", "poisson",coxph", "crr", "boxcox", "ordinal", "geeglm"

offset

string specifying the offset term to be used for Poisson or negative binomial regression. Example: offset="log(follow_up)"

strata

character vector of covariates to stratify by. Only used for coxph and crr

nicenames

boolean indicating if you want to replace . and _ in strings with a space

showN

boolean indicating if you want to show sample sizes

showEvent

boolean indicating if you want to show number of events. Only available for logistic.

CIwidth

width of confidence interval, default is 0.95

reflevel

manual specification of the reference level. Only used for ordinal regression This will allow you to see which model is not fitting if the function throws an error

returnModels

boolean indicating if a list of fitted models should be returned. If this is TRUE then the models will be returned, but the output will be suppressed. In addition to the model elements a data element will be appended to each model so that the fitted data can be examined, if necessary. See Details

fontsize

PDF/HTML output only, manually set the table fontsize

forceWald

forceWald = TRUE is no longer supported; this function will always use profile likelihoods as per the inclusion of the MASS confidence intervals into base from from R 4.4.0

Value

A character vector of the table source code, unless tableOnly=TRUE in which case a data frame is returned

Details

Global p-values are likelihood ratio tests for lm, glm and polr models. For lme models an attempt is made to re-fit the model using ML and if,successful LRT is used to obtain a global p-value. For coxph models the model is re-run without robust variances with and without each variable and a LRT is presented. If unsuccessful a Wald p-value is returned. For GEE and CRR models Wald global p-values are returned.

As of version 0.1.1 if global p-values are requested they will be included in the p-value column.

The number of decimals places to display the statistics can be changed with digits, but this will not change the display of p-values. If more significant digits are required for p-values then use tableOnly=TRUE and format as desired.

tidyselect can only be used for response and covs variables. Additional arguments must be passed in using characters

See also

Examples

# Examples are for demonstration and are not meaningful
# Coxph model with 90% CI
data("pembrolizumab")
rm_uvsum(response = c('os_time','os_status'),
covs=c('age','sex','baseline_ctdna','l_size','change_ctdna_group'),
data=pembrolizumab,CIwidth=.9)
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> HR(90%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Event </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Age at study entry</span> </td>
#>    <td style="text-align:right;"> 0.99 (0.97, 1.00) </td>
#>    <td style="text-align:right;"> 0.16 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Patient Sex</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Female </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 58 </td>
#>    <td style="text-align:right;"> 39 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Male </td>
#>    <td style="text-align:right;"> 1.21 (0.79, 1.84) </td>
#>    <td style="text-align:right;"> 0.46 </td>
#>    <td style="text-align:right;"> 36 </td>
#>    <td style="text-align:right;"> 25 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Baseline ctDNA</span> </td>
#>    <td style="text-align:right;"> 1.00 (1.00, 1.00) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">0.019</span> </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Target lesion size at baseline</span> </td>
#>    <td style="text-align:right;"> 1.00 (1.00, 1.01) </td>
#>    <td style="text-align:right;"> 0.20 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Did ctDNA increase or decrease from baseline to cycle 3</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 73 </td>
#>    <td style="text-align:right;"> 46 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Decrease from baseline </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 33 </td>
#>    <td style="text-align:right;"> 14 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Increase from baseline </td>
#>    <td style="text-align:right;"> 3.06 (1.79, 5.21) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">&lt;0.001</span> </td>
#>    <td style="text-align:right;"> 40 </td>
#>    <td style="text-align:right;"> 32 </td>
#>   </tr>
#> </tbody>
#> </table>

# Linear model with default 95% CI
rm_uvsum(response = 'baseline_ctdna',
covs=c('age','sex','l_size','pdl1','tmb'),
data=pembrolizumab)
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Estimate(95%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Age at study entry</span> </td>
#>    <td style="text-align:right;"> 0.82 (-10.13, 11.76) </td>
#>    <td style="text-align:right;"> 0.88 </td>
#>    <td style="text-align:right;"> 94 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Patient Sex</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 94 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Female </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 58 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Male </td>
#>    <td style="text-align:right;"> 56.61 (-228.71, 341.93) </td>
#>    <td style="text-align:right;"> 0.69 </td>
#>    <td style="text-align:right;"> 36 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Target lesion size at baseline</span> </td>
#>    <td style="text-align:right;"> 1.21 (-1.12, 3.54) </td>
#>    <td style="text-align:right;"> 0.31 </td>
#>    <td style="text-align:right;"> 94 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">PD L1 percent</span> </td>
#>    <td style="text-align:right;"> -3.50 (-8.27, 1.27) </td>
#>    <td style="text-align:right;"> 0.15 </td>
#>    <td style="text-align:right;"> 93 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">log of TMB</span> </td>
#>    <td style="text-align:right;"> 18.78 (-125.18, 162.74) </td>
#>    <td style="text-align:right;"> 0.80 </td>
#>    <td style="text-align:right;"> 94 </td>
#>   </tr>
#> </tbody>
#> </table>

# Logistic model with default 95% CI
rm_uvsum(response = 'os_status',
covs=c('age','sex','l_size','pdl1','tmb'),
data=pembrolizumab,family = binomial)
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> OR(95%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Event </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Age at study entry</span> </td>
#>    <td style="text-align:right;"> 0.98 (0.94, 1.01) </td>
#>    <td style="text-align:right;"> 0.19 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Patient Sex</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Female </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 58 </td>
#>    <td style="text-align:right;"> 39 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Male </td>
#>    <td style="text-align:right;"> 1.11 (0.46, 2.77) </td>
#>    <td style="text-align:right;"> 0.82 </td>
#>    <td style="text-align:right;"> 36 </td>
#>    <td style="text-align:right;"> 25 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Target lesion size at baseline</span> </td>
#>    <td style="text-align:right;"> 1.00 (1.00, 1.01) </td>
#>    <td style="text-align:right;"> 0.37 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">PD L1 percent</span> </td>
#>    <td style="text-align:right;"> 0.98 (0.96, 1.00) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">0.013</span> </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 63 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">log of TMB</span> </td>
#>    <td style="text-align:right;"> 0.70 (0.44, 1.10) </td>
#>    <td style="text-align:right;"> 0.13 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 64 </td>
#>   </tr>
#> </tbody>
#> </table>
# Poisson models returned as model list
mList <- rm_uvsum(response = 'baseline_ctdna',
covs=c('age','sex','l_size','pdl1','tmb'),
data=pembrolizumab, returnModels=TRUE)
#'
# GEE on correlated outcomes
data("ctDNA")
rm_uvsum(response = 'size_change',
covs=c('time','ctdna_status'),
gee=TRUE,
id='id', corstr="exchangeable",
family=gaussian("identity"),
data=ctDNA,showN=TRUE)
#> Warning: Cases with missing response data have been removed.
#>  8 case(s) removed.
#> 8 observations with missing outcome removed.
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Estimate(95%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N (obs|clusters) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Number of weeks on treatment</span> </td>
#>    <td style="text-align:right;"> -0.12 (-0.44, 0.19) </td>
#>    <td style="text-align:right;"> 0.44 </td>
#>    <td style="text-align:right;"> 262|72 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Change in ctDNA since baseline</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 262|72 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Clearance </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 134|14 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> No clearance, decrease from baseline </td>
#>    <td style="text-align:right;"> 61.29 (37.49, 85.09) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">&lt;0.001</span> </td>
#>    <td style="text-align:right;"> 42|17 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> No clearance, increase from baseline </td>
#>    <td style="text-align:right;"> 82.52 (64.75, 100.28) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">&lt;0.001</span> </td>
#>    <td style="text-align:right;"> 86|41 </td>
#>   </tr>
#> </tbody>
#> </table>

# Using tidyselect
pembrolizumab |> rm_uvsum(response = sex,
covs = c(age, cohort))
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: algorithm did not converge
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> <table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> OR(95%CI) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> N </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> Event </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Age at study entry</span> </td>
#>    <td style="text-align:right;"> 1.02 (0.98, 1.05) </td>
#>    <td style="text-align:right;"> 0.38 </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 36 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Study Cohort</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 94 </td>
#>    <td style="text-align:right;"> 36 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> A </td>
#>    <td style="text-align:right;"> Reference </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 16 </td>
#>    <td style="text-align:right;"> 13 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> B </td>
#>    <td style="text-align:right;"> 7.3e-10 (NA, 3.2e+67) </td>
#>    <td style="text-align:right;"> 0.99 </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> C </td>
#>    <td style="text-align:right;"> 7.3e-10 (NA, 3.2e+67) </td>
#>    <td style="text-align:right;"> 0.99 </td>
#>    <td style="text-align:right;"> 18 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> D </td>
#>    <td style="text-align:right;"> 0.16 (0.03, 0.84) </td>
#>    <td style="text-align:right;"> <span style="font-weight: bold;">0.038</span> </td>
#>    <td style="text-align:right;"> 12 </td>
#>    <td style="text-align:right;"> 5 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> E </td>
#>    <td style="text-align:right;"> 0.35 (0.07, 1.36) </td>
#>    <td style="text-align:right;"> 0.15 </td>
#>    <td style="text-align:right;"> 30 </td>
#>    <td style="text-align:right;"> 18 </td>
#>   </tr>
#> </tbody>
#> </table>