Output a scrollable table

This function accepts the output of a aa call to knitr::kable or reportRmd::outTable and, if the output format is html, will produce a scrollable table. Otherwise a regular table will be output for pandoc/latex

scrolling_table(knitrTable, pixelHeight = 500)

Arguments

knitrTable

output from a call to knitr::kable or outTable

pixelHeight

the height of the scroll box in pixels, default is 500

Examples

data("pembrolizumab")
tab <- rm_covsum(data=pembrolizumab,maincov = 'change_ctdna_group',
covs=c('age','cohort','sex','pdl1','tmb','l_size'),full=FALSE)
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
scrolling_table(tab,pixelHeight=300)
#> <div style="border: 1px solid #ddd; padding: 0px; overflow-y: scroll; height:300px;; "><table class="table table" style="margin-left: auto; margin-right: auto; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;position: sticky; top:0; background-color: #FFFFFF;">  </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;position: sticky; top:0; background-color: #FFFFFF;"> Decrease from baseline (n=33) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;position: sticky; top:0; background-color: #FFFFFF;"> Increase from baseline (n=40) </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;position: sticky; top:0; background-color: #FFFFFF;"> p-value </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Age at study entry</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.083 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Mean (sd) </td>
#>    <td style="text-align:right;"> 61.2 (11.6) </td>
#>    <td style="text-align:right;"> 55.2 (14.3) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Median (Min,Max) </td>
#>    <td style="text-align:right;"> 65.1 (34.1, 78.2) </td>
#>    <td style="text-align:right;"> 57.0 (21.1, 81.8) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Study Cohort</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.18 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> A </td>
#>    <td style="text-align:right;"> 6 (18) </td>
#>    <td style="text-align:right;"> 8 (20) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> B </td>
#>    <td style="text-align:right;"> 4 (12) </td>
#>    <td style="text-align:right;"> 7 (17) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> C </td>
#>    <td style="text-align:right;"> 5 (15) </td>
#>    <td style="text-align:right;"> 5 (12) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> D </td>
#>    <td style="text-align:right;"> 8 (24) </td>
#>    <td style="text-align:right;"> 2 (5) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> E </td>
#>    <td style="text-align:right;"> 10 (30) </td>
#>    <td style="text-align:right;"> 18 (45) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Patient Sex</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.84 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Female </td>
#>    <td style="text-align:right;"> 19 (58) </td>
#>    <td style="text-align:right;"> 21 (52) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Male </td>
#>    <td style="text-align:right;"> 14 (42) </td>
#>    <td style="text-align:right;"> 19 (48) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">PD L1 percent</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.26 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Mean (sd) </td>
#>    <td style="text-align:right;"> 23.3 (38.1) </td>
#>    <td style="text-align:right;"> 8.3 (21.8) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Median (Min,Max) </td>
#>    <td style="text-align:right;"> 1 (0, 100) </td>
#>    <td style="text-align:right;"> 0.5 (0.0, 95.0) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">log of TMB</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.42 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Mean (sd) </td>
#>    <td style="text-align:right;"> 1.2 (1.3) </td>
#>    <td style="text-align:right;"> 0.8 (0.7) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Median (Min,Max) </td>
#>    <td style="text-align:right;"> 0.8 (-0.5, 5.2) </td>
#>    <td style="text-align:right;"> 0.7 (-0.8, 3.2) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> <span style="font-weight: bold;">Target lesion size at baseline</span> </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;">  </td>
#>    <td style="text-align:right;"> 0.33 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Mean (sd) </td>
#>    <td style="text-align:right;"> 84.1 (71.0) </td>
#>    <td style="text-align:right;"> 87.4 (49.8) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;padding-left: 2em;" indentlevel="1"> Median (Min,Max) </td>
#>    <td style="text-align:right;"> 64 (11, 387) </td>
#>    <td style="text-align:right;"> 76.5 (15.0, 267.0) </td>
#>    <td style="text-align:right;">  </td>
#>   </tr>
#> </tbody>
#> </table></div>