Set variable labels

Assign variable labels to a data.frame from a lookup table.

set_labels(data, names_labels)

Arguments

data

data frame to be labelled

names_labels

data frame with column 1 containing variable names from data and column 2 containing variable labels. Other columns will be ignored.

Details

Useful if variable labels have been imported from a data dictionary. The first column in names_labels must contain the variable name and the second column the variable label. The column names are not used.

If no label is provided then the existing label will not be changed. To remove a label set the label to NA.

See also

Examples

data("ctDNA")
# create data frame with labels
lbls <- data.frame(c1=c('cohort','size_change'),
c2=c('Cancer cohort','Change in tumour size'))
# set labels and return labelled data frame
set_labels(ctDNA,lbls)
#> # A tibble: 270 × 5
#>    id        cohort ctdna_status                          time size_change
#>    <chr>     <fct>  <fct>                                <dbl>       <dbl>
#>  1 INS-A-002 A      No clearance, increase from baseline  8.12         5.7
#>  2 INS-A-002 A      No clearance, increase from baseline 12.4         NA  
#>  3 INS-A-003 A      No clearance, increase from baseline  6.12        41.7
#>  4 INS-A-007 A      No clearance, decrease from baseline  8.83       -45.5
#>  5 INS-A-007 A      No clearance, decrease from baseline 16.8        -72.7
#>  6 INS-A-007 A      No clearance, decrease from baseline 25.8        -63.6
#>  7 INS-A-007 A      No clearance, decrease from baseline 37.0        -54.5
#>  8 INS-A-007 A      No clearance, decrease from baseline 43.3        -54.5
#>  9 INS-A-007 A      No clearance, decrease from baseline 49.0       -100  
#> 10 INS-A-007 A      No clearance, decrease from baseline 60.0       -100  
#> # ℹ 260 more rows